* Give standard page not found when appropiate
* Unlink the generated files

+++ BartleBlog/backend/web.py
@@ -2,9 +2,10 @@
 
 from colubrid import RegexApplication, HttpResponse, execute
 from colubrid.server import StaticExports
+from colubrid.exceptions import PageNotFound
 from BartleBlog.backend.blog import Blog
 import BartleBlog.backend.dbclasses as db
-import os, codecs
+import os, codecs, sys
 
 class webBlog(Blog):
     def __init__(self):
@@ -18,20 +19,27 @@ class webBlog(Blog):
 class MyApplication(RegexApplication):
     blog=webBlog()
     urls = [
+        ('quit', 'quit'), 
         (r'^(.*?)$', 'page'), 
         (r'^(.*?)/(.*?)$', 'page'), 
         (r'^(.*?)/(.*?)/(.*?)$', 'page'), 
         (r'^(.*?)/(.*?)/(.*?)/(.*?)$', 'page')
     ]
 
+    def quit(self):
+        sys.exit(0)
+
     def page(self, *args):
         path=''.join(args)
         page=db.pageByPath(path)
         self.blog.renderPage(page)
         fn=os.path.join(self.blog.dest_dir, path)
-        resp=HttpResponse(codecs.open(fn).read())
-        os.unlink(fn)
-        return resp
+        if os.path.isfile(fn):
+            resp=HttpResponse(codecs.open(fn).read())
+            os.unlink(fn)
+            return resp
+        else:
+            raise PageNotFound(path)
 
 def run():    
     app = MyApplication
